+++ back/aboutme-adapter/src/main/kotlin/com/aboutme/adapter/common/annotation/AdminController.kt
@@ -1,0 +1,10 @@
+package com.aboutme.adapter.common.annotation
+
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+@RestController
+@RequestMapping("/admin")
+annotation class AdminController
